feat(file-download-helper): create download helper for excel download (#2142)
* feat(file-download-helper): create download helper for excel download


* feat: apply new excel download helpers


* fix: make helper more readable


* fix: remove display, file store(vuex)


* fix: add comments & fix minor things


---------
