+++ apps/web/src/lib/helper/file-download-helper/type.ts
@@ -1,0 +1,33 @@
+import type { DynamicFieldOptions } from '@spaceone/design-system/src/data-display/dynamic/dynamic-field/type/field-schema';
+
+import type { Reference } from '@/lib/reference/type';
+
+interface Options {
+    currency?: string;
+    currencyRates?: Record<string, number>;
+}
+
+export interface ExcelDataField {
+    key: string;
+    name: string;
+    type?: 'datetime'|'enum'|'currency';
+    enum_items?: any;
+    reference?: Reference;
+    options?: DynamicFieldOptions | Options;
+}
+
+interface HeaderMessage {
+    title: string;
+}
+
+export interface ExcelPayload {
+    url?: string;
+    param?: any;
+    data?: any[];
+    fields: ExcelDataField[];
+    sheet_name?: string;
+    header_message?: HeaderMessage;
+    file_name_prefix?: string;
+    version?: string;
+    timezone?: string;
+}
