--- apps/web/src/store/modules/file/getters.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-import { FileState } from '@/store/modules/file/type';
