--- apps/web/src/store/modules/file/mutations.ts
@@ -1,5 +1,0 @@
-import type { FileState } from '@/store/modules/file/type';
-
-export const setDownloadSource = (state: FileState, downloadSource?: string): void => {
-    state.downloadSource = downloadSource;
-};
